requirements.txt
Django
Pillow==9.2.0
python-dotenv==0.21.0
psycopg2-binary==2.9.3
django-crispy-forms==1.14.0
django-formtools==2.3
pycairo==1.23.0
reportlab==4.0.0
